Need to get triangles in there at some point

+++ aux/camera.h
@@ -13,14 +13,14 @@ public:
   Camera();
   virtual ~Camera();
   
-  void setMat(Eigen::MatrixXd m){mat_=m;};
-  Eigen::MatrixXd getMat()const{return mat_;};
+  void setMat(Eigen::MatrixXf m){mat_=m;};
+  Eigen::MatrixXf getMat()const{return mat_;};
   
   Eigen::Vector2f project(const Eigen::Vector3f& v)const;
   Eigen::Vector2f project(const std::vector<float>& v)const;
 
 private:
-  Eigen::MatrixXd mat_;
+  Eigen::MatrixXf mat_;
   // ImageType im_;
 };
